requirements.txt
# Deixando o pip resolver as dependências
streamlit
pandas
langchain-experimental
langchain-groq
matplotlib
tabulate
Pillow
seaborn 